+++ graph/coordinates.h
@@ -7,13 +7,13 @@ using namespace std;
 class coordinates {
 
 public:
-	float x_coordinate, y_coordinate;	
+	double x_coordinate, y_coordinate;	
 	coordinates();
-	coordinates(float x_coordinate, float y_coordinate);
-	float getX_coordinate();
-	float getY_coordinate();
-	void setX_coordinate(float c);
-	void setY_coordinate(float c);
+	coordinates(double x_coordinate, double y_coordinate);
+	double getX_coordinate();
+	double getY_coordinate();
+	void setX_coordinate(double c);
+	void setY_coordinate(double c);
 
 	void printCoordinates(unordered_map<int, coordinates> coor);
 };
